Refactor category model and rename provider to repository
