--- app/src/main/java/app/yournewsontime/data/model/Category.kt
@@ -1,6 +1,0 @@
-package app.yournewsontime.data.model
-
-data class Category(
-    val name: String,
-    var isFollowed: Boolean = false
-)
